Use Azure Pipelines for Windows instead of AppVeyor
This removes the AppVeyor configuration and adapts it for use on Azure
Pipelines. For deployment, the maintainer must create a `docker-creds`
variable group containing `DockerUsername` and `DockerPassword`.
`DockerPassword` should be secret.  This variable group must be linked
to the particular build.

Secrets are not visible to PR builds by default, which mitigates some of
the security risks here. Still, it would be better to opt in to
deployment as opposed to opting out of it with a `condition`.