Export environment from `vcvars64.bat`
Resolves #13.

Some build scripts want to have the compiler and/or linker on the `PATH`
(such as `blt.mond`). On Windows, this is usually achieved by running a
Visual Studio `cmd` shell which sets the appropriate environment.
However, I didn't see a simple way to do this in the `Dockerfile` for
`powershell` (Windows has no equivalent of `source` in `bash`). Instead,
we use a variant of a hack described on [Stack
Overflow](https://stackoverflow.com/a/2124759).

Note the call to `[Environment]::SetEnvironmentVariable`. Simply setting
variables in `$env:` will not persist beyond a single `RUN` command.